refactor: include admin extension into `core`
