+++ lnbits/core/templates/admin/_tab_funding.html
