+++ lnbits/core/views/admin_api.py
@@ -1,34 +1,31 @@
 from http import HTTPStatus
 from typing import Optional
 
-from fastapi import Body, Query
+from fastapi import Body
 from fastapi.params import Depends
 from starlette.exceptions import HTTPException
 
 from lnbits.core.crud import get_wallet
-from lnbits.core.models import User
+from lnbits.core.models import AdminSettings, UpdateSettings, User
+from lnbits.core.services import update_wallet_balance
 from lnbits.decorators import check_admin
-from lnbits.extensions.admin import admin_ext
-from lnbits.extensions.admin.models import AdminSettings, UpdateSettings
 from lnbits.server import server_restart
 
-from .crud import (
-    delete_admin_settings,
-    get_admin_settings,
-    update_admin_settings,
-    update_wallet_balance,
-)
+from .. import core_app
+from ..crud import delete_admin_settings, get_admin_settings, update_admin_settings
 
 
-@admin_ext.get(
-    "/api/v1/restart/", status_code=HTTPStatus.OK, dependencies=[Depends(check_admin)]
+@core_app.get(
+    "/admin/api/v1/restart/",
+    status_code=HTTPStatus.OK,
+    dependencies=[Depends(check_admin)],
 )
 async def api_restart_server() -> dict[str, str]:
     server_restart.set()
     return {"status": "Success"}
 
 
-@admin_ext.get("/api/v1/settings/")
+@core_app.get("/admin/api/v1/settings/")
 async def api_get_settings(
     user: User = Depends(check_admin),  # type: ignore
 ) -> Optional[AdminSettings]:
@@ -36,10 +33,12 @@ async def api_get_settings(
     return admin_settings
 
 
-@admin_ext.put(
-    "/api/v1/topup/", status_code=HTTPStatus.OK, dependencies=[Depends(check_admin)]
+@core_app.put(
+    "/admin/api/v1/topup/",
+    status_code=HTTPStatus.OK,
+    dependencies=[Depends(check_admin)],
 )
-async def api_update_balance(
+async def api_topup_balance(
     id: str = Body(...), amount: int = Body(...)
 ) -> dict[str, str]:
     try:
@@ -54,16 +53,20 @@ async def api_update_balance(
     return {"status": "Success"}
 
 
-@admin_ext.put(
-    "/api/v1/settings/", status_code=HTTPStatus.OK, dependencies=[Depends(check_admin)]
+@core_app.put(
+    "/admin/api/v1/settings/",
+    status_code=HTTPStatus.OK,
+    dependencies=[Depends(check_admin)],
 )
 async def api_update_settings(data: UpdateSettings):
     await update_admin_settings(data)
     return {"status": "Success"}
 
 
-@admin_ext.delete(
-    "/api/v1/settings/", status_code=HTTPStatus.OK, dependencies=[Depends(check_admin)]
+@core_app.delete(
+    "/admin/api/v1/settings/",
+    status_code=HTTPStatus.OK,
+    dependencies=[Depends(check_admin)],
 )
 async def api_delete_settings() -> dict[str, str]:
     await delete_admin_settings()
