--- lnbits/extensions/admin/README.md
@@ -1,12 +1,0 @@
-# Admin Extension
-
-## Dashboard to manage LNbits from the UI
-
-With AdminUI you can manage your LNbits from the UI
-
-![AdminUI](https://i.imgur.com/BIyLkyG.png)
-
-## Before you start
-
-**This extension doesn't discard the need for the `.env` file!**
-In the .env file, set the `LNBITS_ADMIN_USERS` variable to include at least your user id.
