--- lnbits/extensions/admin/config.json
@@ -1,6 +1,0 @@
-{
-	"name": "Admin",
-	"short_description": "Manage your LNbits install",
-	"icon": "build",
-	"contributors": ["benarc"]
-}
